% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.indirect_list.R
\name{as_flextable.indirect_list}
\alias{as_flextable.indirect_list}
\title{Convert an 'indirect_list' Object to a \code{flextable} Object}
\usage{
\method{as_flextable}{indirect_list}(
  x,
  add_sig = FALSE,
  pvalue = FALSE,
  se = TRUE,
  var_labels = NULL,
  digits = 3,
  pval_digits = 3,
  pcut = 0.001,
  use_arrow = TRUE,
  ...
)
}
\arguments{
\item{x}{The object to be converted.
Should be of the class \code{indirect_list}
from the package \code{manymome}.}

\item{add_sig}{Whether a column is
added to denote significant effects
based on confidence intervals. Default
is \code{FALSE}.}

\item{pvalue}{If confidence intervals
are stored, whether asymmetric \emph{p}-values
are reported.}

\item{se}{Whether standard errors
are reported.}

\item{var_labels}{A named vectors.
Used to replace variable names by
other names. For example,
\code{c(x = "I.V", y = "D.V.")} replaces
\code{x} by \code{"I.V"} and \code{y} by \code{"D.V."}
in the output.}

\item{digits}{The number of digits
to be displayed for most columns,
such as indirect effects. Default
is 3.}

\item{pval_digits}{The number of
digits to be displayed for the
\emph{p}-value column, if present. Default
is 3.}

\item{pcut}{Any \emph{p}-value less than
\code{pcut} will be displayed as \verb{p<[pcut]},
\code{"[pcut]"} replaced by the value of
\code{pcut}. Default is .001. Not used
for now.}

\item{use_arrow}{If \code{TRUE}, the
default, use the arrow symbol in
the paths.}

\item{...}{Additional arguments.
Ignored.}
}
\value{
A \code{flextable} object.
}
\description{
The 'as_flextable' method
for the output of 'manymome::many_indirect_effects()'.
}
\details{
It converts an \code{indirect_list}
object,
which is usually crated by
\code{\link[manymome:cond_indirect]{manymome::many_indirect_effects()}},
to a \code{flextable} object. The output
can be further modified by functions
from the package \code{flextable}.
}
\examples{
\donttest{
}

}
